requirements.txt
Django==4.1.7
Pillow==9.4.0
gunicorn==20.1.0
python-dotenv==1.0.0
whitenoise==6.4.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-cors-headers==3.14.0
drf-yasg==1.21.5